BE/ml_models/utils/inference.py
"""
Inference script for license plate detection and recognition
Usage: python inference.py --image <base64_encoded_image>
"""

import sys
import json
import base64
import numpy as np
from PIL import Image
import io
import time

def decode_base64_image(base64_str):
    """
    Decode base64 string to image
    
    Args:
        base64_str: Base64 encoded image string
        
    Returns:
        numpy array: Decoded image
    """
    try:
        img_data = base64.b64decode(base64_str)
        img = Image.open(io.BytesIO(img_data))
        return np.array(img)
    except Exception as e:
        return None

def detect_and_recognize(image):
    """
    Main inference pipeline
    
    Args:
        image: Input image as numpy array
        
    Returns:
        dict: Detection and recognition results
    """
    start_time = time.time()
    
    try:
        # TODO: Load your trained models here
        # detector = PlateDetector('ml_models/plate_detector/model.pt')
        # recognizer = CharacterRecognizer('ml_models/character_recognition/model.pt')
        
        # PLACEHOLDER: Replace with actual model inference
        # This is just example output
        result = {
            'success': True,
            'plate_number': '49P1-12345',  # Replace with actual detection
            'confidence': 0.95,
            'bbox': {
                'x': 100,
                'y': 150,
                'width': 200,
                'height': 50
            },
            'processing_time_ms': int((time.time() - start_time) * 1000)
        }
        
        # Step 1: Detect plate (uncomment when model is ready)
        # plates = detector.detect(image)
        # 
        # if len(plates) == 0:
        #     return {
        #         'success': False,
        #         'message': 'No license plate detected',
        #         'processing_time_ms': int((time.time() - start_time) * 1000)
        #     }
        # 
        # # Get highest confidence plate
        # plate = plates[0]
        # plate_img = plate['image']
        # 
        # # Step 2: Recognize characters
        # plate_text = recognizer.recognize(plate_img)
        # 
        # result = {
        #     'success': True,
        #     'plate_number': plate_text['text'],
        #     'confidence': min(plate['confidence'], plate_text['confidence']),
        #     'bbox': plate['bbox'],
        #     'processing_time_ms': int((time.time() - start_time) * 1000)
        # }
        
        return result
        
    except Exception as e:
        return {
            'success': False,
            'error': str(e),
            'processing_time_ms': int((time.time() - start_time) * 1000)
        }

def main():
    """Main entry point"""
    if len(sys.argv) < 3:
        print(json.dumps({
            'success': False,
            'error': 'Missing image argument'
        }))
        return
    
    # Read image from command line argument
    image_base64 = sys.argv[2]
    
    # Decode image
    image = decode_base64_image(image_base64)
    
    if image is None:
        print(json.dumps({
            'success': False,
            'error': 'Failed to decode image'
        }))
        return
    
    # Run inference
    result = detect_and_recognize(image)
    
    # Print result as JSON
    print(json.dumps(result))

if __name__ == '__main__':
    main()

